Move menuItems and favorite features (get/post/delete) to MenuContext
